requirements.txt
absl-py @ file:///opt/concourse/worker/volumes/live/369745d3-6991-42a3-71cb-795aa3d79471/volume/absl-py_1615411186320/work
aenum==2.2.6
aiohttp==3.7.4
alabaster==0.7.11
amazon-dax-client==2.0.0
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
antlr4-python3-runtime==4.7.2
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astor==0.8.1
astroid==2.0.4
astropy==3.0.4
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
autopep8==2.1.0
aws-xray-sdk==2.9.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt @ file:///opt/concourse/worker/volumes/live/25966884-1a49-4448-62e7-44563eba29d2/volume/bcrypt_1597936171818/work
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
bokeh==0.13.0
boto==2.49.0
boto3==1.18.26
botocore==1.20.112
Bottleneck==1.2.1
certifi==2020.12.5
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.3.9
conda==4.9.2
conda-build==3.15.1
conda-package-handling @ file:///opt/concourse/worker/volumes/live/41a2ba95-d84c-4c06-4d6e-5665866bc5b4/volume/conda-package-handling_1603018173200/work
constantly==15.1.0
contextlib2==0.5.5
coverage @ file:///opt/concourse/worker/volumes/live/4f1a5e30-f675-4496-6570-bc6669a34f2a/volume/coverage_1614614860186/work
cryptography @ file:///opt/concourse/worker/volumes/live/a7af4ad9-2d1a-4e6b-6f87-79f3ea7af2fd/volume/cryptography_1615505226445/work
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distlib==0.3.4
distributed==1.23.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.4.2
Flask==1.1.1
Flask-Cors==3.0.6
fsspec @ file:///tmp/build/80754af9/fsspec_1616582262277/work
future==0.18.2
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
gevent==1.3.6
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
gremlinpython==3.5.1
grpcio @ file:///opt/concourse/worker/volumes/live/4f7ebd3c-7c5e-48f7-6ff5-de127acdc0d8/volume/grpcio_1614884166576/work
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
importlib-metadata @ file:///opt/concourse/worker/volumes/live/ae7e927e-57e2-4966-7b89-25dc6a80092c/volume/importlib-metadata_1615900433179/work
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isodate==0.6.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10.3
jmespath==0.10.0
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
json5==0.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.0
Keras==2.4.3
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown @ file:///opt/concourse/worker/volumes/live/255f8945-2411-4469-7c82-a749af3aef33/volume/markdown_1614363822772/work
MarkupSafe==1.1.1
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mock @ file:///tmp/build/80754af9/mock_1607622725907/work
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multidict==5.1.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
nest-asyncio==1.5.1
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr @ file:///opt/concourse/worker/volumes/live/c0c31ee2-1dfc-42ed-4523-49e57dd0d5ef/volume/numexpr_1614798491918/work
numpy @ file:///opt/concourse/worker/volumes/live/a1b5ea96-b05c-40ba-5dfb-4542ece1c186/volume/numpy_and_numpy_base_1603491228242/work
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.6
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pipenv==2022.1.8
pkginfo==1.4.2
platformdirs==2.4.1
pluggy==0.7.1
ply==3.11
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.14.0
psutil==5.4.7
ptyprocess==0.6.0
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.11.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.5
pyflakes==2.0.0
pygame==2.0.1
Pygments==2.2.0
PyHamcrest @ file:///tmp/build/80754af9/pyhamcrest_1615748656804/work
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyrsistent @ file:///opt/concourse/worker/volumes/live/656e0c1b-ef87-4251-4a51-1290b2351993/volume/pyrsistent_1600141745371/work
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-dotenv==1.0.1
pytz==2018.5
PyWavelets==1.0.0
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
requests==2.19.1
rope==0.11.0
ruamel-yaml==0.15.46
s3transfer==0.5.0
scikit-image==0.14.0
scikit-learn @ file:///opt/concourse/worker/volumes/live/4b416fcf-2a1b-4a30-781a-c588a35f6ca3/volume/scikit-learn_1614446663856/work
scipy @ file:///opt/concourse/worker/volumes/live/ddcf9d39-c185-4171-5824-1a3a0e6d077d/volume/scipy_1614022795557/work
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
slack-bolt==1.18.1
slack-sdk==3.27.1
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
Sphinx==1.7.9
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///home/ktietz/src/ci/sphinxcontrib-htmlhelp_1611920974801/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///home/ktietz/src/ci/sphinxcontrib-serializinghtml_1611920755253/work
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
statsmodels==0.9.0
sympy==1.2
tables==3.4.4
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.5
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tomli==2.0.1
toolz==0.9.0
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
Twisted==19.2.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
unicodecsv==0.14.1
urllib3==1.26.6
virtualenv==20.13.1
virtualenv-clone==0.5.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.1
wrapt==1.10.11
wurlitzer @ file:///opt/concourse/worker/volumes/live/59acb45c-07e7-4304-57af-e84f3889d3ef/volume/wurlitzer_1594753863419/work
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
yarl==1.6.3
zict==0.1.3
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
zope.interface==4.5.0
